Make resolved paths relative to the project or website root
Another approach to Issue #3. Assuming the purpose of a gulp task that includes a resolveUrl step is to produce CSS output usable on the web, the resulting paths should be accessible over HTTP. This approach makes them relative to the HTTP docroot.

This change adds an option to indicate that the paths should be relative to the root. If a valid path is supplied, the path will be calculated relative to it; otherwise if set to true, it will calculate based on the path supplied to root or the current directory. The additional path option is provided for projects where the website's docroot is different from (usually a subdirectory of) the project root -- e.g. composer-based PHP projects where the vendor directory is intentionally placed outside of the HTTP-accessible docroot.

A unit test is included, but some additional coverage is needed to verify the various ways to specify the root.